so much dao architecture stuff

+++ src/dao/postgres/mod.rs
@@ -1,4 +1,5 @@
 pub mod relationships;
+pub mod users;
 
 use crate::config::CONFIG;
 use crate::migration_constants::MIGRATION_FILES;
@@ -6,7 +7,7 @@ use async_trait::async_trait;
 use std::fmt::Debug;
 use std::ops::Deref;
 use tokio_postgres::tls::NoTlsStream;
-use tokio_postgres::{Client, Connection, NoTls, Socket};
+use tokio_postgres::{Client, Connection, NoTls, Row, Socket};
 
 pub async fn apply_migrations(client: Client) {
     let (schema_version, info_exists) = client
@@ -26,24 +27,27 @@ pub async fn apply_migrations(client: Client) {
         .iter()
         .filter(|(i, _)| schema_version.clone() < (i.clone() as i32))
     {
+        println!("{} {}", schema_version, version);
         client.execute(sql.as_str(), &[]).await.unwrap();
         new_schema_version = version.clone() as i32
     }
-    client.execute("DELETE FROM info", &[]).await.unwrap();
-    client
-        .execute(
-            "INSERT INTO info (schema_version) VALUES ($1)",
-            &[&new_schema_version],
-        )
-        .await
-        .unwrap();
+    if new_schema_version.is_positive() {
+        client.execute("DELETE FROM info", &[]).await.unwrap();
+        client
+            .execute(
+                "INSERT INTO info (schema_version) VALUES ($1)",
+                &[&new_schema_version],
+            )
+            .await
+            .unwrap();
+    }
 }
 
 #[async_trait]
-pub trait ConnectionProvider {
-    async fn create_connection() -> (Client, Connection<Socket, NoTlsStream>);
-    async fn open_client() -> Client {
-        let (client, connection) = Self::create_connection().await;
+pub trait ConnectionProvider: std::fmt::Debug {
+    async fn create_connection(&self) -> (Client, Connection<Socket, NoTlsStream>);
+    async fn open_client(&self) -> Client {
+        let (client, connection) = self.create_connection().await;
         tokio::spawn(async move {
             if let Err(e) = connection.await {
                 eprintln!("connection error: {}", e);
@@ -53,22 +57,24 @@ pub trait ConnectionProvider {
     }
 }
 
+#[derive(Debug)]
 pub struct ConfigConnectionProvider;
 
 #[async_trait]
 impl ConnectionProvider for ConfigConnectionProvider {
-    async fn create_connection() -> (Client, Connection<Socket, NoTlsStream>) {
+    async fn create_connection(&self) -> (Client, Connection<Socket, NoTlsStream>) {
         tokio_postgres::connect(CONFIG.deref().db.to_string().as_str(), NoTls)
             .await
             .unwrap()
     }
 }
 
+#[derive(Debug)]
 pub struct DockerConnectionProvider;
 
 #[async_trait]
 impl ConnectionProvider for DockerConnectionProvider {
-    async fn create_connection() -> (Client, Connection<Socket, NoTlsStream>) {
+    async fn create_connection(&self) -> (Client, Connection<Socket, NoTlsStream>) {
         tokio_postgres::connect(
             "user=postgres password=postgres dbname=postgres host=localhost",
             NoTls,
@@ -77,3 +83,25 @@ impl ConnectionProvider for DockerConnectionProvider {
         .unwrap()
     }
 }
+
+pub trait PostgresImpl {
+    fn new(provider: BoxedConnectionProvider) -> Self;
+    fn default() -> Box<Self>
+    where
+        Self: Sized,
+    {
+        if cfg!(test) {
+            Box::new(Self::new(Box::new(DockerConnectionProvider)))
+        } else {
+            Box::new(Self::new(Box::new(ConfigConnectionProvider)))
+        }
+    }
+}
+
+pub trait DbRep {
+    type Output;
+    fn new(row: Row) -> Self;
+    fn model(self) -> Self::Output;
+}
+
+pub type BoxedConnectionProvider = Box<dyn ConnectionProvider + Sync + std::marker::Send>;
